rename package from clocks to prci
