+++ src/main/scala/prci/ClockNodes.scala
@@ -1,5 +1,5 @@
 // See LICENSE.SiFive for license details.
-package freechips.rocketchip.clocks
+package freechips.rocketchip.prci
 
 import chisel3._
 import chisel3.internal.sourceinfo.SourceInfo
