+++ src/main/scala/prci/ResetWrangler.scala
@@ -1,5 +1,5 @@
 // See LICENSE for license details.
-package freechips.rocketchip.clocks
+package freechips.rocketchip.prci
 
 import chisel3._
 import chisel3.util._
